# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto
  description: |
    Keep in touch with your friends
    by sharing photos of special moments,
    thanks to WASAPhoto! You can
    upload your photos directly from your PC,
    and they will be visible to everyone following you.
  version: 1.0.0
security:
  - bearerAuth: []
...
paths:
...
  /following/{id}/:
    parameters:
      - $ref: "#/components/parameters/user_id"

    put:
      operationId: followUser
      tags: [following]
      summary: Follows a user
      description: |-
        The user follows another user
      responses: 
        #No Content
        '204':
          description: The user has been followed successfully
        #BadRequest
        '400':
          description: Invalid operatation, maybe invalid user id
        #Unauthorized
        '401':
          $ref: '#/components/responses/Unauthorized'
        #Forbidden
        '403':
          $ref: '#/components/responses/Forbidden'
        #InternalServerError
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      operationId: unfollowUser
      tags: [following]
      summary: Unfollows a user
      description: |-
        The user unfollows another user
      responses: 
        #No Content
        '204':
          description: The user has been unfollowed successfully
        #BadRequest
        '400':
          description: Invalid operatation, maybe invalid user id
        #Unauthorized
        '401':
          $ref: '#/components/responses/Unauthorized'
        #InternalServerError
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /photos/:
    post:
      operationId: uploadPhoto
      tags: [photo]
      summary: Uploads a photo
      description: |-
        The user uploads a photo
      requestBody:
        description: binary image file
        required: true
        content:
          image/*:
            schema: 
              $ref: "#/components/schemas/BinaryImage"
      responses:
        #CREATED
        '201':
          description: The photo has been created successfully
          content: 
            application/json:
              schema: { $ref: '#/components/schemas/photo'}
        #BadRequest
        '400':
          description: The request was not valid. probably Invalid photo formatting
        #Unauthorized
        '401':
          $ref: '#/components/responses/Unauthorized'
        #InternalServerError
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /likes/{pid}/:
  
    parameters:
      - $ref: "#/components/parameters/pid"
    
    put:
      operationId: likePhoto
      tags: [like]
      summary: Likes a photo
      description: |-
        The user likes a photo
      responses:
        #No Content
        '204':
          description: The photo has been liked successfully
        #BadRequest
        '400':
          description: Invalid operation, maybe invalid photo id
        #Unauthorized
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        #InternalServerError
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      operationId: unlikePhoto
      tags: [like]
      summary: Unlikes a photo
      description: |-
        The user unlikes a photo
      responses:
        #No Content
        '204':
          description: The photo has been unliked successfully
        #BadRequest
        '400':
          description: Invalid operation, maybe invalid photo id
        #Unauthorized
        '401':
          $ref: '#/components/responses/Unauthorized'
        #InternalServerError
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /photos/{pid}/comment/{comment_id}/:
      
        parameters:
          - $ref: "#/components/parameters/pid"
          - $ref: "#/components/parameters/comment_id"
        
        delete:
          operationId: uncommentPhoto
          tags: [comment]
          summary: uncomment a comment
          description: |-
            The user uncomment a comment
          responses:
            #No Content
            '204':
              description: The comment has been deleted successfully
            #BadRequest
            '400':
              description: The request was not valid. probably Invalid comment id
            #Unauthorized
            '401':
              $ref: '#/components/responses/Unauthorized'
            #Forbidden
            '403':
              $ref: '#/components/responses/Forbidden'
            #InternalServerError
            '500':
              $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
...
    Unauthorized:
      description: The request was not authorized. The user must log in.

    #403
    Forbidden:
      description: The request was not authorized. The user does not have the necessary permissions.

    #500
    InternalServerError:
      description: An error occurred on the server. Please try again later.
   

###################PAMETERS####################
  parameters:
    user_id:
      name: id
      in: path
      description:  it's a user id
      required: true
      schema:
        $ref: "#/components/schemas/user_id"
    comment_id:
      name: comment_id
      in: path
      description: it's a comment's id
      required: true
      schema:
        $ref: "#/components/schemas/comment"
    pid:
      name: pid
      in: path
      description: it's a photo's id
      required: true
      schema:
        $ref: "#/components/schemas/pid"
    

      
        

###################SCHEMAS####################
  schemas:
    user_id:
      description: String between 3-16 characters
      type: string
      pattern: '^.*?$'
      example: MickeyMouse
      minLength: 3
      maxLength: 16
      readOnly: true
    username:
      title: username
      description: "This object represents the name of the user"
      type: string
      pattern: '^.*?$'
      example: MickeyMouse
      minLength: 3
      maxLength: 16
...
    pid:
      title: pid
      description: "This object represents the id of the photo"
      type: number
      format: int64
      example: 1234
      readOnly: true
    comment_id:
      title: comment_id
      description: "This object represents the id of the comment"
      type: number
      format: int64
      example: 1234
      readOnly: true
...
    photo:
      title: photo
      description: "This object represents a single picture/photo"
      type: object
      properties:
        user_id:
          $ref: "#/components/schemas/user_id"
        username:
          $ref: "#/components/schemas/username"
        pid:
          $ref: "#/components/schemas/pid"
        likes:
          description: "users that liked the photo"
          type: array
          minItems: 0
          maxItems: 100
          items: 
            $ref: "#/components/schemas/user_id"
        comments:
          description: this is all the comments of this photo
          type: array
          minItems: 0
          maxItems: 100
          items:  
            $ref: "#/components/schemas/comment"
        date:
          description: Date and time of upload.
          type: string
          format: date-time
          example: 2019-02-21T17:32:28Z
          readOnly: true
          minLength: 1
          maxLength: 2200
    text:
      description: Comment text.
      type: string
      pattern: '^.*?$'
      minLength: 1
      maxLength: 2200
      example: "I love this your dog!"
    comment:
      title: comment
      description: "This object represents a single comment"
      properties:
        owner:
          $ref: "#/components/schemas/user_id"
        surname:
          $ref: "#/components/schemas/username"
        pid:
          $ref: "#/components/schemas/pid"
        comment_id:
          $ref : "#/components/schemas/comment_id"
        text:
          $ref: "#/components/schemas/text"
        date:
          description: Date and time of upload.
          type: string
          format: date-time
          example: 2019-02-21T17:32:28Z
          minLength: 1
          maxLength: 2200
          readOnly: true
    BinaryImage:
      description: Binary image.
      type: string
      format: binary
      minLength: 4
      maxLength: 65535
      example: iVBORw0KGgoAAAANSU... 
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer